// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "movies".
 */
export interface Movie {
  id: number;
  name: string;
  url: string;
  votes: number;
  poster: number | Media;
  overview: string;
  tagline: string;
  genres: {
    name?: string | null;
    id?: string | null;
  }[];
  slug?: string | null;
  updatedAt: string;
  createdAt: string;
}